Extract new env group form to a new independant component
